# zip-repl/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GEF Player Statistics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a new team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    put:
      operationId: updateTeam
      tags: [teams]
      summary: Update team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /upload/image:
    post:
      operationId: uploadImage
      tags: [upload]
      summary: Upload player image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    UploadResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        teamId:
          type: integer
          nullable: true
        teamName:
          type: string
          nullable: true
        nationality:
          type: string
          nullable: true
        marketValue:
          type: number
          nullable: true
        salary:
          type: number
          nullable: true
        matchesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        goalsScored:
          type: integer
        goalsConceded:
          type: integer
        mvpCount:
          type: integer
        overallRating:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - matchesPlayed
        - wins
        - losses
        - draws
        - goalsScored
        - goalsConceded
        - mvpCount
        - overallRating
        - createdAt
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        createdAt:
          type: string
      required:
        - id
        - name
        - players
        - createdAt
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
      required:
        - name
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
...